tests/lib/nested.sh: also configure chrony if used

+++ tests/lib/nested.sh
@@ -1499,15 +1499,21 @@ nested_setup_vm(){
         remote.exec "sudo sync"
     fi
     if [ -n "${NTP_SERVER:-}" ]; then
-        # Configure systemd-timesyncd to use the predefined ntp server
-        CONF_FILE="/etc/systemd/timesyncd.conf"
-        remote.exec "cp \"$CONF_FILE\" /tmp/timesyncd.conf"
-        remote.exec "sed -i -e '/^NTP=/d' -e '/^FallbackNTP=/d' /tmp/timesyncd.conf"
-        remote.exec "sed -i '/^\[Time\]/a NTP='\"$NTP_SERVER\" /tmp/timesyncd.conf"
-        remote.exec "sed -i '/^\[Time\]/a FallbackNTP=' /tmp/timesyncd.conf"
-        remote.exec "sudo cp /tmp/timesyncd.conf \"$CONF_FILE\""
-        remote.exec "sudo systemctl restart systemd-timesyncd"
-        remote.exec "sudo sync"
+        if remote.exec "[ -d /etc/chrony/sources.d ]"; then
+            remote.exec "sudo rm /etc/chrony/sources.d/*.sources"
+            echo "pool ${NTP_SERVER} iburst maxsources 1 nts prefer" | remote.exec "sudo tee /etc/chrony/sources.d/proxy.sources"
+            remote.exec "sudo systemctl try-restart chrony.service"
+        fi
+        if remote.exec "[ -f /etc/systemd/timesyncd.conf ]"; then
+            # Configure systemd-timesyncd to use the predefined ntp server
+            CONF_FILE="/etc/systemd/timesyncd.conf"
+            remote.exec "cp \"$CONF_FILE\" /tmp/timesyncd.conf"
+            remote.exec "sed -i -e '/^NTP=/d' -e '/^FallbackNTP=/d' /tmp/timesyncd.conf"
+            remote.exec "sed -i '/^\[Time\]/a NTP='\"$NTP_SERVER\" /tmp/timesyncd.conf"
+            remote.exec "sed -i '/^\[Time\]/a FallbackNTP=' /tmp/timesyncd.conf"
+            remote.exec "sudo cp /tmp/timesyncd.conf \"$CONF_FILE\""
+            remote.exec "sudo systemctl try-restart systemd-timesyncd"
+        fi
     fi
 }
